combine CromwellConfig tests into passthrough tests
